More state_dict tests
